part-2/application/requirements.txt
Flask==3.0.2
gunicorn==21.2.0
azure-storage-blob==12.19.0